watcher/requirements.txt
Flask==1.0.2
kafka-python==1.4.6
